+++ app/Http/Requests/UpdateEmployeeRequest.php
@@ -201,6 +201,7 @@ class UpdateEmployeeRequest extends FormRequest
                         ->where('legal_entity_id', $legalEntityId);
                 })->whereNull('deleted_at'),
             ],
+            'organizational_unit_id' => ['prohibited'],
         ], $this->employeeAddressItemRules(), $this->certificationValidationRules(true));
     }